refactor: change auth file name
