+++ lib/adapters/auth/auth_adapter.dart
