+++ lib/ui/auth/login/view_model/login_view_model.dart
@@ -1,4 +1,4 @@
-import 'package:finance_ai/adapters/auth/storage_adapter.dart';
+import 'package:finance_ai/adapters/auth/auth_adapter.dart';
 import 'package:finance_ai/utils/command.dart';
 import 'package:finance_ai/utils/result.dart';
 import 'package:logging/logging.dart';
